requirements.txt
python-dotenv
discord.py
sense_hat
requests